requirements.txt
pandas
gspread
google-auth-oauthlib
playwright